chore(deps)!: update boto3, botocore, and testing dependencies
Updating dependencies. Notably:
boto3 to at least 1.28.78
botocore is added as a new explicit dep

Testing dependencies are also updated for:
pytest-timeout to 2.2.* from 2.1.*
black to 23.10.* from 23.9.*
mypy to 1.6.* from 1.5.1
ruff to 0.1.* from 0.0.288

*BREAKING CHANGE* - This is only a breaking change because of how the
deadline-cloud and deadline-cloud-worker-agent currently define their
boto3 dependency as "boto3 ~= 1.26", so if we were to release this as
a patch release then they would be unable to resolve their test
dependencies.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>